Continued work on Sparkle Fizz
The issue is that for max acceleration greater than MAX_G_SPARKLEFIZZ,
it takes a while for acc_decayed to be equal to MAX_G_SPARKLEFIZZ. If
acc_max is twice MAX_G_SPARKLEFIZZ, half of ACC_MAX_DECAY_RATE is spent
at solid sparkle, leaving only the other half to get down to zero (for
a linear decay).

What’s needed is a separate decay timescale that starts at that moment,
maybe using an if statement to choose how acc_decayed is calculated:
linear at first, then when acc_decayed is below MAX_G_SPARKLEFIZZ, take
a new t0 and decay according to this new decay rate.

This decay should probably be exponential but slow. A sigmoid might
work well too.