Created Physical version of DHO
Finally revisited the DHO and acceleration equations. This allowed me
to use the acceleration value in m/sˆ2 and ultimately express the ball
position as a value centered on 0.0 when upright and -2.0 when upside
down—exactly as one would expect for a mass on a spring at rest at x =
0.0.

This also allows me to change m, k and c without having to rescale
everything.

Transforming ball position to strand position was trivial.