Corrected COLOR_ORDER, added Fire With Palette
Used RGBCalibrate example to finally figure out what the order should
be. Amazing I could get it wrong a few times and still have the colors
I wanted. I guess all of Flamingo was pink, which kinda worked, then
green, which was wrong in the first place, and rainbows. And white. Wow.