Added some notes on power: the battery connector is not likely to be useful
Since if USB is unplugged, the only route to the LED VCC (VBUS) is through the 5 V regulator. This is also true if a battery backpack is used.

Does it make sense to put the batt connector on the VUSB line? Not really, because then the Arduino just sees unregulated batt... which is OK for a LiPo I guess.

Does it make sense to change the LEDs to use VBAT? With VBAT and no USB present, the LEDS get direct battery and the arduino gets regulated output. With USB and no VBAT the LEDs dont see VBUS unless a battery backpack is used.